fix pipelines via changing to service
