+++ build.gradle
@@ -29,7 +29,7 @@ ext {
 }
 
 group = "stirling.software"
-version = "0.45.5"
+version = "0.45.6"
 
 java {
     // 17 is lowest but we support and recommend 21
