+++ src/main/resources/templates/fragments/common.html
@@ -240,8 +240,8 @@
         window.stirlingPDF.sessionExpired = /*[[#{session.expired}]]*/ '';
         window.stirlingPDF.refreshPage = /*[[#{session.refreshPage}]]*/ 'Refresh Page';
         window.stirlingPDF.error = /*[[#{error}]]*/ "Error";
-        window.stirlingPDF.uploadLimit = /*[[${uploadLimit}]]*/ 0;
-        window.stirlingPDF.uploadLimitReadable = /*[[${uploadLimitReadable}]]*/ 'Unlimited';
+        window.stirlingPDF.uploadLimitReadable = /*[[${@uploadLimitService.getReadableUploadLimit()}]]*/ 'Unlimited';
+        window.stirlingPDF.uploadLimit = /*[[${@uploadLimitService.getUploadLimit()}]]*/ 0;
         window.stirlingPDF.uploadLimitExceededSingular = /*[[#{uploadLimitExceededSingular}]]*/ 'is too large. Maximum allowed size is';
         window.stirlingPDF.uploadLimitExceededPlural = /*[[#{uploadLimitExceededPlural}]]*/ 'are too large. Maximum allowed size is';
       })();
@@ -292,10 +292,10 @@
       </div>
     </div>
     <div class="selected-files flex-wrap"></div>
-    <div class="text-muted small mt-0 text-end w-100" th:if="${uploadLimit != 0}">
-      <span th:text="#{uploadLimit}">Maximum file size: </span>
-      <span th:text="${uploadLimitReadable}"></span>
-    </div>
+	<div class="text-muted small mt-0 text-end w-100" th:if="${@uploadLimitService.getUploadLimit() != 0}">
+	    <span th:text="#{uploadLimit}">Maximum file size: </span>
+	    <span th:text="${@uploadLimitService.getReadableUploadLimit()}"></span>
+	</div>
   </div>
   <div class="progressBarContainer" style="display: none; position: relative;">
     <div class="progress" style="height: 1rem;">
